feat: Implemented Branch Table Commands to Rush Cargo Insiders
* Fully Implemented Branch Table and Branch Classes
* Fully Implemented Branch rm, get, all, mod and add Commands
* Split 'events.py' Module into 'events.py', 'buildingEvents.py' and 'territoryEvents.py' Modules
* Fixed Minor Bugs
