+++ rushcargo-insiders/src/lib/model/database_building.py
@@ -1,16 +1,14 @@
 from psycopg import sql
 
 from .classes import Building, Warehouse, Branch
-
 from .constants import *
-
+from .database import Database, console
 from .database_tables import (
     SpecializationTable,
     noCoincidenceFetched,
     insertedRow,
     getTable,
 )
-from .database import Database, console
 
 
 # Functions that Returns Some Generic Table-related Strings
@@ -32,7 +30,7 @@ class BuildingTable(SpecializationTable):
         )
 
     # Returns Building Insert Query
-    def __insertQuery(self):
+    def __insertParentQuery(self):
         return sql.SQL(
             "INSERT INTO {parentTableName} ({fields}) VALUES (%s,%s, %s, %s, %s, %s, %s)"
         ).format(
@@ -50,13 +48,6 @@ class BuildingTable(SpecializationTable):
             ),
         )
 
-    # Filter Items with Multiple Conditions from Building Table
-    def _getMult(self, fields: list[str], values: list) -> bool:
-        if not SpecializationTable._getMultParentTable(self, fields, values):
-            return False
-
-        return True
-
     # Find Building from Building Table
     def _find(self, areaId: int, buildingName: str) -> Building | None:
         """
@@ -64,8 +55,8 @@ class BuildingTable(SpecializationTable):
         """
 
         # Get Building
-        if not self._getMult(
-            [BUILDING_FK_CITY_AREA, BUILDING_NAME], [areaId, buildingName]
+        if not SpecializationTable._getMultParentTable(
+            self, [BUILDING_FK_CITY_AREA, BUILDING_NAME], [areaId, buildingName]
         ):
             return None
 
@@ -73,9 +64,9 @@ class BuildingTable(SpecializationTable):
         return Building.fromItemFetched(self._items[0])
 
     # Insert Building to Building Table
-    def _add(self, b: Building) -> None:
+    def _add(self, b: Building | Warehouse | Branch) -> None:
         # Get Query to Insert Building to Building Table
-        query = self.__insertQuery()
+        query = self.__insertParentQuery()
 
         # Execute Query
         try:
@@ -119,11 +110,6 @@ class WarehouseTable(BuildingTable):
         # Number of Items
         nItems = len(self._items)
 
-        # No Results
-        if nItems == 0:
-            noCoincidenceFetched()
-            return
-
         # Initialize Rich Table
         table = getTable("Warehouse", nItems)
 
@@ -131,10 +117,9 @@ class WarehouseTable(BuildingTable):
         table.add_column("ID", justify="left", max_width=ID_NCHAR)
         table.add_column("Name", justify="left", max_width=LOCATION_NAME_NCHAR)
         table.add_column("Coords", justify="left", max_width=COORDINATE_NCHAR)
-        table.add_column("Description", justify="left", max_width=DESCRIPTION_NCHAR)
         table.add_column("Phone", justify="left", max_width=CONTACT_NCHAR)
         table.add_column("Email", justify="left", max_width=CONTACT_NCHAR)
-        table.add_column("City Area ID", justify="left", max_width=ID_NCHAR)
+        table.add_column("City Area", justify="left", max_width=ID_NCHAR)
 
         # Loop Over Items
         for item in self._items:
@@ -146,7 +131,6 @@ class WarehouseTable(BuildingTable):
                 str(w.buildingId),
                 w.buildingName,
                 getCoords(w.gpsLatitude, w.gpsLongitude),
-                w.addressDescription,
                 str(w.phone),
                 w.email,
                 str(w.areaId),
@@ -167,14 +151,11 @@ class WarehouseTable(BuildingTable):
         # Get Warehouse Insert Query
         warehouseQuery = self.__insertQuery()
 
-        # Get Building object from Warehouse
-        b = Building.fromWarehouse(w)
-
         # Insert Building to Building Table
-        BuildingTable._add(self, b)
+        BuildingTable._add(self, w)
 
         # Get Building ID
-        b = BuildingTable._find(self, b.areaId, b.buildingName)
+        b = BuildingTable._find(self, w.areaId, w.buildingName)
 
         # Execute Query to Insert Warehouse
         try:
@@ -187,19 +168,34 @@ class WarehouseTable(BuildingTable):
                 insertedRow(w.buildingName, self._tableName),
                 style="success",
             )
-            
+
         except Exception as err:
             raise err
 
     # Filter Items from Warehouse Table
     def get(self, field: str, value, printItems: bool = True) -> bool:
         if not SpecializationTable._getTable(self, field, value):
+            if printItems:
+                noCoincidenceFetched()
             return False
 
         # Print Items
         if printItems:
             self.__print()
         return True
+
+    # Find Warehouse from Warehouse Table
+    def find(self, warehouseId: int) -> Warehouse | None:
+        """
+        Returns Warehouse Object if it was Found. Otherwise, False
+        """
+
+        # Get Warehouse
+        if not SpecializationTable._getTable(self, WAREHOUSE_ID, warehouseId):
+            return None
+
+        # Get Warehouse Object from Item Fetched
+        return Warehouse.fromItemFetched(self._items[0])
 
     # Get All Items from Warehouse Table
     def all(self, orderBy: str, desc: bool) -> None:
@@ -235,15 +231,10 @@ class BranchTable(BuildingTable):
 
     # Print Items
     def __print(self) -> None:
-        w = None
+        b = None
 
         # Number of Items
         nItems = len(self._items)
-
-        # No Results
-        if nItems == 0:
-            noCoincidenceFetched()
-            return
 
         # Initialize Rich Table
         table = getTable("Warehouse", nItems)
@@ -252,69 +243,76 @@ class BranchTable(BuildingTable):
         table.add_column("ID", justify="left", max_width=ID_NCHAR)
         table.add_column("Name", justify="left", max_width=LOCATION_NAME_NCHAR)
         table.add_column("Coords", justify="left", max_width=COORDINATE_NCHAR)
-        table.add_column("Description", justify="left", max_width=DESCRIPTION_NCHAR)
         table.add_column("Phone", justify="left", max_width=CONTACT_NCHAR)
         table.add_column("Email", justify="left", max_width=CONTACT_NCHAR)
+        table.add_column("Warehouse ID", justify="left", max_width=ID_NCHAR)
+        table.add_column("Distance", justify="left", max_width=DISTANCE_NCHAR)
         table.add_column("City Area ID", justify="left", max_width=ID_NCHAR)
 
         # Loop Over Items
         for item in self._items:
-            # Intialize Warehouse from Item Fetched
-            w = Warehouse.fromItemFetched(item)
+            # Intialize Branch from Item Fetched
+            b = Branch.fromItemFetched(item)
 
             # Add Row to Rich Table
             table.add_row(
-                str(w.buildingId),
-                w.buildingName,
-                getCoords(w.gpsLatitude, w.gpsLongitude),
-                w.addressDescription,
-                str(w.phone),
-                w.email,
-                str(w.areaId),
+                str(b.buildingId),
+                b.buildingName,
+                getCoords(b.gpsLatitude, b.gpsLongitude),
+                str(b.phone),
+                b.email,
+                str(b.warehouseConnection),
+                str(b.routeDistance),
+                str(b.areaId),
             )
 
         # Print Table
         console.print(table)
 
-    # Returns Warehouse Insert Query
+    # Returns Branch Insert Query
     def __insertQuery(self):
-        return sql.SQL("INSERT INTO {tableName} ({field}) VALUES (%s)").format(
+        return sql.SQL("INSERT INTO {tableName} ({fields}) VALUES (%s, %s, %s)").format(
             tableName=sql.Identifier(self._tableName),
-            field=sql.Identifier(self._tablePKFKName),
+            fields=sql.SQL(",").join(
+                [
+                    sql.Identifier(self._tablePKFKName),
+                    sql.Identifier(BRANCH_FK_WAREHOUSE_CONNECTION),
+                    sql.Identifier(BRANCH_ROUTE_DISTANCE),
+                ]
+            ),
         )
 
-    # Insert Warehouse to Warehouse Table
-    def add(self, w: Warehouse) -> None:
-        # Get Warehouse Insert Query
-        warehouseQuery = self.__insertQuery()
-
-        # Get Building object from Warehouse
-        b = Building.fromWarehouse(w)
+    # Insert Branch to Branch Table
+    def add(self, b: Branch) -> None:
+        # Get Branch Insert Query
+        branchQuery = self.__insertQuery()
 
         # Insert Building to Building Table
         BuildingTable._add(self, b)
 
         # Get Building ID
-        b = BuildingTable._find(self, b.areaId, b.buildingName)
+        B = BuildingTable._find(self, b.areaId, b.buildingName)
 
         # Execute Query to Insert Warehouse
         try:
             self._c.execute(
-                warehouseQuery,
-                [b.buildingId],
+                branchQuery,
+                [B.buildingId, b.warehouseConnection, b.routeDistance],
             )
 
             console.print(
-                insertedRow(w.buildingName, self._tableName),
+                insertedRow(b.buildingName, self._tableName),
                 style="success",
             )
 
         except Exception as err:
             raise err
 
-    # Filter Items from Warehouse Table
+    # Filter Items from Branch Table
     def get(self, field: str, value, printItems: bool = True) -> bool:
         if not SpecializationTable._getTable(self, field, value):
+            if printItems:
+                noCoincidenceFetched()
             return False
 
         # Print Items
@@ -322,22 +320,36 @@ class BranchTable(BuildingTable):
             self.__print()
         return True
 
-    # Get All Items from Warehouse Table
+    # Find Branch from Branch Table
+    def find(self, branchId: int) -> Branch | None:
+        """
+        Returns Branch Object if it was Found. Otherwise, False
+        """
+
+        # Get Warehouse
+        if not SpecializationTable._getTable(self, BRANCH_ID, branchId):
+            return None
+
+        # Get Branch Object from Item Fetched
+        return Branch.fromItemFetched(self._items[0])
+
+    # Get All Items from Branch Table
     def all(self, orderBy: str, desc: bool) -> None:
         SpecializationTable._all(self, orderBy, desc)
 
         # Print Items
         self.__print()
 
-    # Modify Row from Warehouse Table
-    def modify(self, warehouseId: int, field: str, value) -> None:
-        """
-        NOTE: There's No Own Warehouse Field that can be Modified, Only the Ones Inherited from its Parent Table
-        """
+    # Modify Row from Branch Table
+    def modify(self, branchId: int, field: str, value) -> None:
+        # Modify Branch Table Row Columns
+        if field == BRANCH_FK_WAREHOUSE_CONNECTION or field == BRANCH_ROUTE_DISTANCE:
+            SpecializationTable._modifyTable(self, branchId, field, value)
 
         # Modify Building Table Row Column
-        SpecializationTable._modifyParentTable(self, warehouseId, field, value)
+        else:
+            SpecializationTable._modifyParentTable(self, branchId, field, value)
 
-    # Remove Row from Warehouse Table
-    def remove(self, warehouseId: int) -> None:
-        SpecializationTable._remove(self, warehouseId)
+    # Remove Row from Branch Table
+    def remove(self, branchId: int) -> None:
+        SpecializationTable._remove(self, branchId)
